requirements.txt
streamlit==1.57.0
langchain==1.3.0
langgraph==1.2.0
langchain-google-genai==4.2.2
tavily-python==0.7.24
python-dotenv==1.2.2